backend/requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.5
atpublic==5.0
attrs==24.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cronsim==2.6
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dj-database-url==3.0.1
Django==5.2.3
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.7.0
django-stubs-ext==5.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.7
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
fido2==1.2.0
filelock==3.16.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
fonttools==4.55.0
fqdn==1.5.1
fsspec==2025.7.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.3
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.1.3
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numpy==1.26.4
oncalendar==1.1
openai==1.98.0
openpyxl==3.1.5
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.1.3
pandocfilters==1.5.1
parso==0.8.4
Pillow==10.0.0
platformdirs==4.3.6
playwright==1.54.0
plotly==5.17.0
prometheus_client==0.22.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.3
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pycurl==7.45.4
pydantic==2.10.3
pydantic_core==2.27.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.8.0
pymongo==3.11.4
pyotp==2.9.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==3.3.0
python-magic-bin==0.4.14
pytz==2024.2
pywin32==310
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.4.0
rcssmin==1.1.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rjsmin==1.2.2
rpds-py==0.25.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
segno==1.6.1
selenium==4.34.0
Send2Trash==1.8.3
sentence-transformers==5.0.0
setuptools==68.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsd==4.0.1
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.21.4
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.4.0
urlopen==1.0.0
uvicorn==0.34.0
virtualenv==20.28.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.40.0
whitenoise==6.8.2
word2number==1.1
wsproto==1.2.0
